feat: upgrade kill processes script

+++ Shell/killProcesses.sh
@@ -4,11 +4,11 @@
 # take an array of ports to kill processes by arguments from command line
 # example: kp 8080 cypress 9229 3000 node 
 ARG_ARRAY=("$@")
-PORTS=()
-
 
 ARG_PORTS=($(printf "%s\n" "${ARG_ARRAY[@]}" | grep -E '^[0-9]+$'))
-ARG_PROCESSES=($(printf "%s\n" "${ARG_ARRAY[@]}" | grep -E '^[a-zA-Z]+$'))
+
+# to filter out the processes by names: support letters and hyphens
+ARG_PROCESSES=($(printf "%s\n" "${ARG_ARRAY[@]}" | grep -E '^[a-zA-Z-]+$'))
 
 if [ ${#ARG_PORTS[@]} -eq 0 ] && [ ${#ARG_PROCESSES[@]} -eq 0 ]; then
     echo "No ports or processes provided. Exiting."
@@ -16,55 +16,35 @@ if [ ${#ARG_PORTS[@]} -eq 0 ] && [ ${#ARG_PROCESSES[@]} -eq 0 ]; then
 fi
 
 if [ ${#ARG_PORTS[@]} -gt 0 ]; then
+    echo PROCESSES TO BY KILLED BY PORTS: ${ARG_PORTS[@]}
     for PORT in "${ARG_PORTS[@]}"; do
-        PORTS+=($PORT)
+        PID=$(lsof -t -i:$PORT)
+        if [ -n "$PID" ]; then
+            kill -9 $PID
+            echo "Killed process of by ports of PID: $PID"
+        else
+            echo "No process found by port $PORT"
+        fi
     done
-    echo PORTS TO BE KILLED: ${PORTS[@]}
 fi
+
 
 if [ ${#ARG_PROCESSES[@]} -gt 0 ]; then
-    echo PROCESSES TO BE KILLED: ${ARG_PROCESSES[@]}
-fi
-
-
-for PORT in "${PORTS[@]}"; do
-    PID=$(lsof -t -i:$PORT)
-    if [ -n "$PID" ]; then
-        kill -9 $PID
-        echo "Killed process $PID on port $PORT"
-    else
-        echo "No process found on port $PORT"
-    fi
-done
-
-if [[ " ${ARG_PROCESSES[@]} " =~ " cypress " ]]; then
-    CY_PIDs=$(ps aux | grep cypress | awk '{print $2}')
-    for CY_PID in $CY_PIDs; do
-        kill -9 $CY_PID
-        echo "Killed cypress processes of PIDs: $CY_PID"
+    echo PROCESSES TO BE KILLED BY NAMES: ${ARG_PROCESSES[@]}
+    for PROCESS in "${ARG_PROCESSES[@]}"; do
+        echo "Trying to kill the process by names: $PROCESS"
+        PIDs=$(pgrep $PROCESS)
+        if [ -z "$PIDs" ]; then
+            echo "No process found by name: $PROCESS"
+        else
+            echo "PIDs found by names: $PIDs"
+            for PID in $PIDs; do
+                if [ -n "$PID" ]; then 
+                    kill -9 $PID
+                    echo "Killed processes by names of PID: $PID"
+                fi
+            done
+        fi
     done
 fi
 
-if [[ " ${ARG_PROCESSES[@]} " =~ " gradle " ]]; then
-    GRADLE_PIDs=$(ps aux | grep gradle | awk '{print $2}')
-    for GRADLE_PID in $GRADLE_PIDs; do
-        kill -9 $GRADLE_PID
-        echo "Killed Gradle process of pid: $GRADLE_PID"
-    done
-fi
-
-if [[ " ${ARG_PROCESSES[@]} " =~ " node " ]]; then
-    NODE_PIDs=$(ps aux | grep node | awk '{print $2}')
-    for NODE_PID in $NODE_PIDs; do
-        kill -9 $NODE_PID
-        echo "Killed node process of pid: $NODE_PID"
-    done
-fi
-
-if [[ " ${ARG_PROCESSES[@]} " =~ " sshagent " ]]; then
-    SSHAGENTs=$(ps aux | grep ssh-agent | awk '{print $2}')
-    for SSHAGENT in $SSHAGENTs; do
-        kill -9 $SSHAGENT
-        echo "Killed ssh-agent process of pid: $SSHAGENT"
-    done
-fi
